[JAMES-3836] Extracts both mail repository loaders to their own module
GuiceMailRepositoryLoader is the original implementation which creates a loader with a fake module to inject the url before requesting the instanciation of the MailRepository
MultiBoundGuiceRepositoryLoader retrieves a multi binder (aka a Set) of MailRepositoryFactories, maps them into a dictionnary of factories indexed by the instanciated type and delegates the actual instance creation to the factories. This way there is no need to bind a method parameter in a subcontext only to make the container do the injection :)
