+++ server/container/guice/mailrepository-loaders/src/main/java/org/apache/james/modules/mailrepository/guice/MultiBoundGuiceRepositoryLoader.java
@@ -17,37 +17,33 @@
  * under the License.                                           *
  ****************************************************************/
 
-package org.apache.james.modules.server;
+package org.apache.james.modules.mailrepository.guice;
+
+import java.util.Map;
+import java.util.Optional;
+import java.util.Set;
+import java.util.function.Function;
+import java.util.stream.Collectors;
+
+import javax.inject.Inject;
 
 import org.apache.james.mailrepository.api.MailRepository;
+import org.apache.james.mailrepository.api.MailRepositoryFactory;
 import org.apache.james.mailrepository.api.MailRepositoryLoader;
 import org.apache.james.mailrepository.api.MailRepositoryStore;
 import org.apache.james.mailrepository.api.MailRepositoryUrl;
-import org.apache.james.utils.ClassName;
-import org.apache.james.utils.GuiceGenericLoader;
 
-import com.google.inject.Inject;
-import com.google.inject.Module;
-
-public class GuiceMailRepositoryLoader implements MailRepositoryLoader {
-    private final GuiceGenericLoader genericLoader;
+public class MultiBoundGuiceRepositoryLoader implements MailRepositoryLoader {
+    private final Map<String, Function<MailRepositoryUrl, MailRepository>> factoriesByType;
 
     @Inject
-    public GuiceMailRepositoryLoader(GuiceGenericLoader genericLoader) {
-        this.genericLoader = genericLoader;
+    private MultiBoundGuiceRepositoryLoader(Set<MailRepositoryFactory> factories) {
+        factoriesByType = factories.stream().collect(Collectors.toMap(factory -> factory.mailRepositoryClass().getName(), it -> it::create));
     }
 
-    @Override
     public MailRepository load(String fullyQualifiedClassName, MailRepositoryUrl url) throws MailRepositoryStore.MailRepositoryStoreException {
-        try {
-            Module urlModule = binder -> binder.bind(MailRepositoryUrl.class).toInstance(url);
-
-            return genericLoader.<MailRepository>withChildModule(urlModule)
-                .instantiate(new ClassName(fullyQualifiedClassName));
-        } catch (ClassNotFoundException e) {
-            throw new MailRepositoryStore.UnsupportedRepositoryStoreException("No Mail Repository found with class name " + fullyQualifiedClassName);
-        } catch (ClassCastException e) {
-            throw new MailRepositoryStore.MailRepositoryStoreException(fullyQualifiedClassName + " is not a MailRepository");
-        }
+        return Optional.ofNullable(factoriesByType.get(fullyQualifiedClassName))
+                .map(factory -> factory.apply(url))
+                .orElseThrow(() -> new MailRepositoryStore.MailRepositoryStoreException("no factory for " + fullyQualifiedClassName));
     }
 }
