+++ server/container/guice/pom.xml
@@ -51,6 +51,7 @@
         <module>mailbox-plugin-deleted-messages-vault</module>
         <module>mailet</module>
         <module>mailrepository-cassandra</module>
+        <module>mailrepository-loaders</module>
         <module>memory</module>
         <module>onami</module>
         <module>opensearch</module>
@@ -252,6 +253,11 @@
             <dependency>
                 <groupId>${james.groupId}</groupId>
                 <artifactId>mailrepository-cassandra</artifactId>
+                <version>${project.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>${james.groupId}</groupId>
+                <artifactId>mailrepository-loaders</artifactId>
                 <version>${project.version}</version>
             </dependency>
             <dependency>
